requirements.txt
# Core ML Libraries (para uso local do modelo)
torch>=2.0.0
transformers>=4.57.3
peft>=0.18.0
trl>=0.25.1
datasets>=4.4.1
tokenizers>=0.22.1

# API Client (obrigatório para uso com API)
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
accelerate>=0.30.0
safetensors>=0.4.0
numpy>=1.24.0
pandas>=2.0.0

# Optional but recommended
bitsandbytes>=0.41.0  # Para quantização 8-bit/4-bit
scipy>=1.11.0

